# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ideas": {
      "post": {
        "operationId": "IdeasController_createIdea",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIdeaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Ideas"
        ]
      },
      "get": {
        "operationId": "IdeasController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ideas"
        ]
      }
    },
...
    "/ideas/{id}": {
      "get": {
        "operationId": "IdeasController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ideas"
        ]
      },
      "delete": {
        "operationId": "IdeasController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ideas"
        ]
      }
    },
...
  },
  "info": {
    "title": "WebAstro API",
    "description": "API para la Agencia Digital WebAstro",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateIdeaDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}